pyproject.toml
[project]
name = "pdf-to-excel-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "playwright>=1.53.0",
    "pyotp>=2.9.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
]
